Move WeaklyCanonicalPath AppContainer fallback into the Windows port
Per repo convention, platform-specific code belongs under core/platform/<port>/
behind a cross-platform Env interface, not inline in core/framework/ behind
#ifdef _WIN32. This commit moves the AppContainer fallback there:

* Add Env::GetWeaklyCanonicalPath as a pure virtual.
* Implement WindowsEnv::GetWeaklyCanonicalPath in core/platform/windows/env.cc
  (where WindowsEnv::GetCanonicalPath already lives). The AppContainer-specific
  ERROR_ACCESS_DENIED fallback that uses GetFinalPathNameByHandleW with
  VOLUME_NAME_NT now lives entirely in this port.
* Implement PosixEnv::GetWeaklyCanonicalPath in core/platform/posix/env.cc
  (where PosixEnv::GetCanonicalPath already lives) as a thin
  std::filesystem::weakly_canonical wrapper.
* tensorprotoutils.cc::WeaklyCanonicalPath becomes a six-line shim that calls
  Env::Default().GetWeaklyCanonicalPath. No #ifdef _WIN32, no Win32 includes,
  no platform helpers in framework code.
* The Windows test seam used to verify the fallback algorithm directly moves
  from utils::internal in tensorprotoutils.h to onnxruntime::internal in
  core/platform/windows/env.h, next to its implementation.

Behavior is unchanged: non-Windows builds and non-AppContainer Windows runs
still hit std::filesystem::weakly_canonical first, the AppContainer fallback
still only fires on ERROR_ACCESS_DENIED, and all 13 PathValidationTest cases
still pass.
